requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
protobuf==4.21.10
pydub==0.25.1
scipy==1.9.3
soundfile==0.11.0
librosa==0.9.2
torch==1.13.0
numpy==1.23.5